requirements.txt
hopsworks
jupyter
pandas
pyarrow
tqdm
plotly
scikit-learn
xgboost
optuna
python-dotenv
matplotlib
chardet
charset-normalizer
numpy
streamlit
geopandas
lightgbm
python-dotenv